DT-778 Add (very) bvasic UI test
