+++ src/main/kotlin/uk/gov/justice/hmpps/offenderevents/service/OffenderEventStore.kt
@@ -17,4 +17,8 @@ class OffenderEventStore {
     return store.toList()
   }
 
+  fun clear() {
+    store.clear()
+  }
+
 }
